# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "API de Transferências",
    "version": "1.0.0",
    "description": "API para login, registro, consulta de usuários e transferências."
  },
  "paths": {
    "/users/register": {
      "post": {
        "summary": "Registrar usuário",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "username": { "type": "string" },
                  "password": { "type": "string" },
                  "favorecido": { "type": "boolean" }
                },
                "required": ["username", "password"]
              }
            }
          }
        },
        "responses": {
          "201": { "description": "Usuário registrado" },
          "409": { "description": "Usuário já existe" }
        }
      }
    },
    "/users/login": {
      "post": {
        "summary": "Login de usuário",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "username": { "type": "string" },
                  "password": { "type": "string" }
                },
                "required": ["username", "password"]
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Login bem-sucedido" },
          "401": { "description": "Credenciais inválidas" }
        }
      }
    },
    "/users": {
      "get": {
        "summary": "Consultar usuários",
        "responses": {
          "200": { "description": "Lista de usuários" }
        }
      }
    },
    "/transfers": {
      "post": {
        "summary": "Realizar transferência",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "from": { "type": "string" },
                  "to": { "type": "string" },
                  "value": { "type": "number" }
                },
                "required": ["from", "to", "value"]
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Transferência realizada" },
          "400": { "description": "Erro na transferência" }
        }
      },
      "get": {
        "summary": "Consultar transferências",
        "responses": {
          "200": { "description": "Lista de transferências" }
        }
      }
    }
  }
}